ucm-mva/requirements.txt
pytest~=7.4.4
igraph~=0.11.6
Pillow~=10.4.0
geopy~=2.4.1
numpy~=1.23.5
rasterio~=1.3.11
matplotlib~=3.9.2
scipy~=1.14.1
saas @ git+https://github.com/cooling-singapore/saas-middleware@8654f3892e49c854dbe056560bfb02226e05a059
